requirements.txt
# Core dependencies for Retail Sales Insight Pipeline
flask>=3.0.0
pandas>=2.0.0
PyYAML>=6.0
pytest>=7.0.0
requests>=2.31.0

# Oracle database connection (if needed)
cx_Oracle==8.3

# Development and testing
python-dotenv>=1.0.0
